Move Gaussian band model to HGPS catalog code
